Revert "Made some modifications to the footer"
