+++ backend/app/auth.py
@@ -21,7 +21,7 @@ ALGORITHM = "HS256"
 
 
 # Create an instance of the OAuth
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl = 'users/login')
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl = 'login')
 
 # Exception for invalid credentials
 credentials_exception = HTTPException(
