+++ backend/app/emails.py
@@ -1,6 +1,6 @@
 from fastapi import (BackgroundTasks, UploadFile,File, Form, Depends, HTTPException, status)
-from fastapi_mail import FastMail, ConnectionConfig, MessageSchema, MessageType
-from typing import List, Dict, Any
+from fastapi_mail import FastMail, ConnectionConfig, MessageSchema
+from typing import List
 from jose import jwt, JWTError
 from fastapi.exceptions import HTTPException
 from datetime import datetime, timedelta
@@ -13,8 +13,6 @@ from crud import get_user_by_email
 from sqlalchemy.orm import Session
 from dotenv import load_dotenv
 import os
-from pydantic import EmailStr, BaseModel
-
 
 
 # Load all environment variables
@@ -30,14 +28,8 @@ conf = ConnectionConfig(
     MAIL_STARTTLS = False,
     USE_CREDENTIALS = True,
     MAIL_SSL_TLS= True,
-    VALIDATE_CERTS = True,
-    TEMPLATE_FOLDER='../templates'
-
+    VALIDATE_CERTS = True
 )
-
-class EmailSchema(BaseModel):
-    body: Dict[str, Any]
-    
 
 async def send_email(email: List, instance: User):
     token_data = {
@@ -46,33 +38,28 @@ async def send_email(email: List, instance: User):
     }
 
     token = jwt.encode(token_data, os.getenv('SECRET'), algorithm='HS256')
-    
-    emails: EmailSchema = {
-        "body": {
-            "url": f"https://api.heed.hng.tech/verification?token={token}"
-        } 
-    }
 
-    # template = f"""
-    #     <div>
-    #                 <h3>Account Verification </h3>
-    #                 <br>
-    #                 <p>Thank you for registering with us. Kindly click on the link below to
-    #                 verify your email and have full acccess to the platform.</p>
 
-    #                 <a href="https://api.heed.hng.tech/verification?token={token}">Verify your email address </a>
-    #     </div>
-    # """
+    template = f"""
+        <div>
+                    <h3>Account Verification </h3>
+                    <br>
+                    <p>Thank you for registering with us. Kindly click on the link below to
+                    verify your email and have full acccess to the platform.</p>
+
+                    <a href="https://heed.hng.tech/emailverification?token={token}">Verify your email address </a>
+        </div>
+    """
 
     message = MessageSchema(
         subject = "Account Verification",
         recipients =email,
-        template_body=emails.get("body"),
-        subtype=MessageType.html,
+        body = template,
+        subtype = "html"
     )
 
     fm =FastMail(conf)
-    await fm.send_message(message=message, template_name='EmailVerification/index.html')
+    await fm.send_message(message=message)
 
 
 async def verify_token(token: str, db: Session):
