+++ backend/app/jwt.py
@@ -7,8 +7,6 @@ from pydantic import BaseModel
 from sqlalchemy.orm import Session
 import models
 from crud import get_user_by_email
-from fastapi.responses import JSONResponse
-from fastapi.encoders import jsonable_encoder
 # from db import engine, SessionLocal
 
 from dotenv import load_dotenv
@@ -42,6 +40,9 @@ user_not_found_exception = HTTPException(
     detail="User not found",
     headers={"WWW-Authenticate": "Bearer"},
 )
+
+
+
 
 # Pydantic classes for type checking
 class Token(BaseModel):
@@ -122,37 +123,31 @@ async def main_login(form_data, db):
     N.B: form_data is a parameter in your endpoint should depend on OAuth2PasswordRequestForm, i.e:
             form_data: OAuth2PasswordRequestForm = Depends()
     """
+
+
+    user = authenticate_user(db, form_data.username, form_data.password)
+    if not user:
+        raise HTTPException(
+            status_code=400,
+            detail="Incorrect email or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
     
-    try:
-        user = authenticate_user(db, form_data.username, form_data.password)
-        if not user:
-            raise HTTPException(
-                status_code=400,
-                detail="Incorrect email or password",
-                headers={"WWW-Authenticate": "Bearer"},
-            )
-        
-        if not user.is_active:
-            raise HTTPException(
-                status_code=400,
-                detail=f"Sorry {user.first_name}, your account is yet to be activated.",
-                headers={"WWW-Authenticate": "Bearer"},
-            )
-        
-        access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
-        refresh_token_expires = timedelta(minutes=REFRESH_TOKEN_EXPIRE_MINUTES)
-        access_token = create_access_token(
-            data={"sub": user.email}, expires_delta=access_token_expires
+    if not user.is_active:
+        raise HTTPException(
+            status_code=400,
+            detail=f"Sorry {user.first_name}, your account is yet to be activated.",
+            headers={"WWW-Authenticate": "Bearer"},
         )
-        refresh_token = create_refresh_token(
-            data={"sub": user.email}, expires_delta=refresh_token_expires
-        )
-        
-    except Exception as e:
-        return JSONResponse(
-            status_code= status.HTTP_400_BAD_REQUEST,
-            content=jsonable_encoder({"detail": str(e)}),
-        )
+    
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    refresh_token_expires = timedelta(minutes=REFRESH_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(
+        data={"sub": user.email}, expires_delta=access_token_expires
+    )
+    refresh_token = create_refresh_token(
+        data={"sub": user.email}, expires_delta=refresh_token_expires
+    )
     return {"access_token": access_token,
             "refresh_token": refresh_token,
             "token_type": "bearer"}
