+++ backend/app/requirements.txt
@@ -13,7 +13,6 @@ colorama==0.4.6
 dnspython==2.2.1
 ecdsa==0.18.0
 email-validator==1.3.0
-email-validate==1.1.2
 fastapi==0.87.0
 fastapi-mail==1.2.1
 fastapi-pagination==0.10.0
