+++ backend/app/schema.py
@@ -192,9 +192,3 @@ class ChangePassword(BaseModel):
 
 class RefreshToken(BaseModel):
     refresh_token:str
-
-class Newsletter(BaseModel):
-    email: EmailStr
-
-    class Config:
-        orm_mode = True
