+++ backend/templates/EmailVerification/index.css
@@ -1,0 +1,5 @@
+*{
+    box-sizing: border-box;
+    padding: 0;
+    border: 0;
+}
