+++ backend/templates/EmailVerification/index.html
@@ -1,107 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Heed</title>
-
-    <style>
-      @import url("https://fonts.googleapis.com/css2?family=Kanit:wght@300;400;600;800&display=swap");
-      * {
-        box-sizing: border-box;
-        padding: 0;
-        border: 0;
-      }
-      .main {
-        padding-inline: 40px;
-        max-width: 1440px;
-        width: 100%;
-        padding-top: 22px;
-        font-family: "Kanit", sans-serif;
-        color: #3f3f3f;
-      }
-      .message {
-        display: block;
-        text-align: center;
-      }
-      .logo img {
-        width: 150px;
-        max-width: 30%;
-      }
-      .heed {
-        color: #006cff;
-      }
-      .mail {
-        width: 100%;
-        max-width: 300px;
-      }
-      p {
-        font-size: 14px;
-        font-weight: 300;
-        line-height: 24px;
-        text-align: center;
-      }
-      .copyright {
-        text-align: center;
-      }
-      .copyright p {
-        font-weight: 400;
-      }
-      .message a {
-        color: white;
-        text-decoration: none;
-        background-color: #006cff;
-        color: #ffffff;
-        padding-inline: 61.5px !important;
-        padding-block: 16px !important;
-        font-family: inherit;
-        font-size: 16px;
-        font-weight: 500;
-        border-radius: 4px;
-        margin-left: auto;
-        padding: 16px 20px !important;
-      }
-      .email {
-        margin-top: 30px;
-        margin-bottom: 20px;
-      }
-      .copy {
-        font-size: 16px;
-        padding-inline: 5px;
-      }
-    </style>
-  </head>
-  <body>
-    <div class="main">
-      <div class="logo">
-        <a href="https://heed.cx"
-          ><img
-            src="https://res.cloudinary.com/dcsv5etuw/image/upload/v1671352986/logo_rwgpkh.jpg"
-            alt=""
-        /></a>
-      </div>
-      <div class="message">
-        <img
-          src="https://res.cloudinary.com/dcsv5etuw/image/upload/v1671353020/mail_p2jchn.jpg"
-          alt=""
-          class="mail"
-        />
-        <h1>Welcome to <span class="heed">Heed!</span></h1>
-        <p>
-          You registered an account on Heed, before being able to use your
-          account you need to verify that this is your email address by clicking
-          the button below:
-        </p>
-        <div class="email">
-          <a href="www.google.com">Verify email address</a>
-        </div>
-        <p>If you have any questions, please Contact Us</p>
-      </div>
-      <div class="copyright">
-        <!-- <img src="https://res.cloudinary.com/dcsv5etuw/image/upload/v1671353006/copyright_frjemi.jpg" alt=""> -->
-        <p><span class="copy">©</span>Copyright 2022 | Heed</p>
-      </div>
-    </div>
-  </body>
+    <link rel="stylesheet" href="index.css">
+</head>
+<body>
+    <h1>Hello Heed</h1>
+</body>
 </html>
