+++ scrybe_frontend/src/components/Footer/footer.module.scss
@@ -41,7 +41,6 @@
     }
     .joinContainer {
       display: flex;
-      padding-top: 20px;
       //       margin-top: 31vh;
       p {
         font-weight: 400;
@@ -98,13 +97,6 @@
         font-weight: 400;
         font-size: 0.9em;
         line-height: 24px;
-      }
-      a{
-        color:white;
-        text-align:center
-      }
-      a:hover{
-        color:#006cff
       }
       .footer__right__logo {
         padding-right: 5%;
