+++ scrybe_frontend/src/components/Footer/index.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { BsInstagram, BsLinkedin, BsTwitter } from "react-icons/bs";
+import { BsInstagram, BsTwitter } from "react-icons/bs";
 import { GrFacebookOption } from "react-icons/gr";
 import { Link, NavLink } from "react-router-dom";
 import locate from "./assets/location-marker.png";
@@ -10,7 +10,6 @@ function Footer() {
   React.useEffect(() => {
     // 👇️ scroll to top on page load
     window.scrollTo({ top: 0, left: 0 });
-    // window.open('mailto:email@example.com?subject=Subject&body=Body%20goes%20here')
   }, []);
   return (
     <div className={styles.footer}>
@@ -35,13 +34,13 @@ function Footer() {
 
             <div className={styles.joinContainer}>
               <div className={styles.socials}>
-                <a href="https://www.linkedin.com/company/heed-cx/" target="_blank">
-                  <BsLinkedin />
+                <a href="https://heed.hng.tech">
+                  <BsInstagram />
                 </a>
-                <a href="https://www.facebook.com/profile.php?id=100088581892757&mibextid=ZbWKwL" target="_blank">
+                <a href="https://heed.hng.tech">
                   <GrFacebookOption />
                 </a>
-                <a href="https://twitter.com/heed_cx" target="_blank">
+                <a href="https://heed.hng.tech">
                   <BsTwitter />
                 </a>
               </div>
@@ -51,6 +50,11 @@ function Footer() {
           <div className={styles.footerRight}>
             <ul>
               <h4> Explore</h4>
+              <li>
+                <NavLink to="/" className={styles.links}>
+                  Home
+                </NavLink>
+              </li>
               <li>
                 <NavLink to="/about-us" className={styles.links}>
                   About Us
@@ -72,7 +76,7 @@ function Footer() {
               <h4>Quick links</h4>
               <li>
                 <NavLink to="/pricing" className={styles.links}>
-                  Pricing
+                  Billing and pricing
                 </NavLink>
               </li>
               <li>
@@ -82,7 +86,7 @@ function Footer() {
               </li>
               <li>
                 <NavLink to="/help-support" className={styles.links}>
-                  Support
+                  Help and Support
                 </NavLink>
               </li>
             </ul>
@@ -90,16 +94,22 @@ function Footer() {
               <h4>Contact</h4>
               <div className={styles.footer__right__contact}>
                 <div className={styles.footer__right__logo}>
-                  <img src={email} alt="emailpic" />
+                  <img src={call} alt="" />
                 </div>
-                <a href="mailto:support@heed.cx">info@heed.cx </a>
-              </div> 
+                <p>+2348790900976</p>
+              </div>
               <div className={styles.footer__right__contact}>
                 <div className={styles.footer__right__logo}>
-                  <img src={email} alt="emailpic" style={{maxWidth:20 , height : 20}} />
+                  <img src={email} alt="" />
                 </div>
-                <a href="mailto:support@heed.cx">support@heed.cx </a>
-              </div> 
+                <p>info@heed.co.za</p>
+              </div>
+              <div className={styles.footer__right__contact}>
+                <div className={styles.footer__right__logo}>
+                  <img src={locate} alt="" />
+                </div>
+                <p>54 jones close Abuja</p>
+              </div>
             </div>
           </div>
         </div>
